clientes

+++ src/main/java/com/uniamerica/aluguelEquipamento/controller/ClientesController.java
@@ -1,6 +1,4 @@
 package com.uniamerica.aluguelEquipamento.controller;
-
-
 import com.uniamerica.aluguelEquipamento.model.Clientes;
 import com.uniamerica.aluguelEquipamento.service.ClientesService;
 import org.springframework.http.HttpStatus;
@@ -18,12 +16,17 @@ public class ClientesController {
         this.clientesService = clientesService;
     }
 
-    @GetMapping(path = "/{nome}")
-    public ResponseEntity<Clientes> findByEmail(@PathVariable String nome) {
-        return ResponseEntity.ok(clientesService.findByNome(nome));
+    @GetMapping(path = "/nome/{nome}")
+    public ResponseEntity<?> findByEmail(@PathVariable String nome) {
+        List<Clientes> list = clientesService.findByNome(nome);
+
+        if(!list.isEmpty()){
+            return new ResponseEntity<>(list, null, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(list, null, HttpStatus.NO_CONTENT);
     }
 
-    @GetMapping(path = "/{id}")
+    @GetMapping(path = "/id/{id}")
     public ResponseEntity<Clientes> findById(@PathVariable Long id) {
         return ResponseEntity.ok(clientesService.findById(id));
     }
@@ -57,15 +60,28 @@ public class ClientesController {
     }
 
     @DeleteMapping(path = "/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        clientesService.delete(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<Void> delete(@PathVariable long id) {
+
+        Clientes clientes1 = clientesService.findById(id);
+
+        if(clientes1 != null){
+            clientesService.delete(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Clientes> replace(@RequestBody Clientes clientes,
-                                            @PathVariable Long id) {
-        clientesService.update(clientes, id);
-        return new ResponseEntity<>(HttpStatus.OK);
+                                            @PathVariable long id) {
+        Clientes clientes1 = clientesService.findById(id);
+
+        if(clientes1 != null){
+            clientesService.update(clientes, id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
     }
 }
